+++ src/Temperature/Temperature.cpp
@@ -2,12 +2,12 @@
 
 namespace devuino
 {
-    float Temperature::fahrenheit()
+    double Temperature::fahrenheit()
     {
         return celsius() * (9.0 / 5.0) + 32.0;
     }
 
-    float Temperature::kelvin()
+    double Temperature::kelvin()
     {
         return celsius() + 273.15;
     }
